vote-interface: move vote state v3 to new module
